fix for missing lib
git-svn-id: svn://svn.dd-wrt.com/DD-WRT@28956 52c4871e-980c-0410-b1e0-e73912ce01f8

Former-commit-id: 0ce35cdbadd8f381bb35236be9eced312cebbe78

+++ src/router/rules/asterisk.mk
@@ -206,7 +206,9 @@ asterisk-install:
 	rm -rf $(INSTALLDIR)/util-linux/bin
 	rm -rf $(INSTALLDIR)/util-linux/sbin
 	rm -f $(INSTALLDIR)/util-linux/lib/libmount.so*
+ifneq ($(CONFIG_E2FSPROGS_ADV),y)
 	rm -f $(INSTALLDIR)/util-linux/lib/libblkid.so*
+endif
 	rm -f $(INSTALLDIR)/util-linux/lib/libfdisk.so*
 	rm -f $(INSTALLDIR)/util-linux/lib/libsmartcols.so*
